+++ extensions/anytype/src/components/UpdateForm/UpdateTagForm.tsx
@@ -32,11 +32,11 @@ export function UpdateTagForm({ spaceId, propertyId, tag, mutateTags }: UpdateTa
         });
 
         await updateTag(spaceId, propertyId, tag.id, {
-          name: values.name || "",
+          name: values.name,
           color: values.color as Color,
         });
 
-        showToast(Toast.Style.Success, "Tag updated successfully");
+        await showToast(Toast.Style.Success, "Tag updated successfully");
         mutateTags();
         pop();
       } catch (error) {
@@ -60,8 +60,8 @@ export function UpdateTagForm({ spaceId, propertyId, tag, mutateTags }: UpdateTa
         </ActionPanel>
       }
     >
-      <Form.TextField {...itemProps.name} title="Name" placeholder="Enter tag name" info="The name of the tag" />
-      <Form.Dropdown {...itemProps.color} title="Color" info="The color of the tag">
+      <Form.TextField {...itemProps.name} title="Name" placeholder="Add name" info="The name of the tag" />
+      <Form.Dropdown {...itemProps.color} title="Color" placeholder="Select color" info="The color of the tag">
         {tagColorKeys.map((key) => {
           const value = Color[key];
           return (